+++ src/main/java/org/folio/linked/data/mapper/dto/resource/common/title/ParallelTitleMapperUnit.java
@@ -8,7 +8,6 @@ import static org.folio.ld.dictionary.PropertyDictionary.PART_NAME;
 import static org.folio.ld.dictionary.PropertyDictionary.PART_NUMBER;
 import static org.folio.ld.dictionary.PropertyDictionary.SUBTITLE;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.PARALLEL_TITLE;
-import static org.folio.linked.data.util.ResourceUtils.getFirstValue;
 import static org.folio.linked.data.util.ResourceUtils.putProperty;
 
 import com.fasterxml.jackson.databind.JsonNode;
@@ -51,7 +50,7 @@ public class ParallelTitleMapperUnit extends TitleMapperUnit {
   public Resource toEntity(Object dto, Resource parentEntity) {
     var parallelTitle = ((ParallelTitleField) dto).getParallelTitle();
     var resource = new Resource();
-    resource.setLabel(getLabel(getFirstValue(parallelTitle::getMainTitle), getFirstValue(parallelTitle::getSubTitle)));
+    resource.setLabel(getLabel(parallelTitle));
     resource.addTypes(PARALLEL_TITLE);
     resource.setDoc(getDoc(parallelTitle));
     resource.setIdAndRefreshEdges(hashService.hash(resource));